pyproject.toml
[project]
name = "aws-cdk-apigw-demo"
version = "0.1.0"
description = "AWS CDK API Gateway Demo"
readme = "README.md"
requires-python = ">=3.11.7"
dependencies = [
    "aws-lambda-powertools>=3.5.0",
    "pydantic>=2.10.6",
]

[dependency-groups]
dev = [
    "aws-cdk-aws-lambda-python-alpha>=2.178.0a0",
    "aws-cdk-lib>=2.178.0",
    "ipython>=8.32.0",
    "pytest>=8.3.4",
]
